Use `WidgetKey` to test using driver and without using `material.dart`
So, it turns out that when integration testing, the flutter app runs
using flutter support package on the app. But then, the test code are
run on the local machine, which then using `FlutterDriver` APIs to
interact with the on-device app code (using the observatory-port)

As a reason of this, the integration test's code cannot include any
flutter specific code as it is only available on the device.